requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiohttp-jinja2==1.6
aiosignal==1.3.2
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
click==8.2.1
exceptiongroup==1.3.0
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jinja2==3.1.6
linkify-it-py==2.0.3
markdown-it-py==3.0.0
markupsafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
msgpack==1.1.0
multidict==6.4.4
platformdirs==4.3.8
propcache==0.3.1
pygments==2.19.1
rich==14.0.0
sniffio==1.3.1
textual==3.2.0
textual-dev==1.7.0
textual-serve==1.1.2
tqdm==4.67.1
typing-extensions==4.13.2
uc-micro-py==1.0.3
yarl==1.20.0
discord.py==2.3.2
python-dotenv==1.0.1
